bug fixing
git-svn-id: svn+ssh://cdcvs.fnal.gov/cvs/projects/damic/test/extract@87 678ac9d6-6348-4232-9a94-0a0f05251285

+++ gConfig.h
@@ -19,7 +19,7 @@ class gConfig
 	
 	float getSeedThr()   {return fSeedThr;};
 	float getAddThr()    {return fAddThr;};
-	float getSkirtSize() {return fSkirtSize;};
+	int   getSkirtSize() {return fSkirtSize;};
 	
 	bool  getSaveTracks(){return fSaveTracks;};
 	std::string getTracksCuts(){return fTracksCuts;};
